// generated/src/app/api-models/cap-fund-dtl.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CapFundDtl {
  capFundModelId: string;
  capFundSubModelId: string;
  capFundModelDesc: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}